Add files via upload
Create, Read/Retrieve, Update, and Delete are the four steps in the JSP CRUD process. The four basic roles of persistence storage are as follows.

The JSP CRUD operation can be defined as user interface conventions that allow the view, search, and modify information through computer-based forms and reports. JSP CRUD is a data-driven framework that makes use of HTTP action verbs in a consistent manner. There are a few key verbs in HTTP.

POST: This method creates a new resource.
GET: Retrieves information from a resource.
PUT: Replaces/updates an existing resource with a new one.
DELETE: Removes a resource from the system.

Each of these procedures corresponds to a set of commands in a database